cityAlert/requirements.txt
Django==1.6.4
Pillow==2.4.0
South==0.8.4
psycopg2==2.5.2
python-dateutil==2.2
python-mimeparse==0.1.4
django-tastypie==0.11.0
django-qsstats-magic==0.7.2
